backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
pandas==2.2.1
spotipy==2.23.0
python-dotenv==1.0.1
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
typing_extensions==4.9.0
tzdata==2024.1 